+++ kindle2notion/parsing.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Tuple
 
 from dateparser import parse
 
-from kindle2notion.parsing.WordDetector import WordDetector
+from kindle2notion.parsing.word_detector import WordDetector
 from kindle2notion.parsing.enums import NoteEnum, PageEnum, LocationEnum, DateAddedEnum, Locale, Word
 
 BOOKS_WO_AUTHORS = []
